Move handlers into their own module
